modify socket functions param type from int to size_t

+++ utils/socket.c
@@ -52,7 +52,7 @@ int vcrypto_connect(char* socket_file_path) {
   return conn_fd;
 }
 
-bool vcrypto_recvmsg(int connfd, void* recv_data_buf, int recv_len, int *recv_fd, int num_fd) {
+bool vcrypto_recvmsg(int connfd, void* recv_data_buf, size_t recv_len, int *recv_fd, int num_fd) {
   if (num_fd != 1 && num_fd != 0) {
     log_error("vcrypto_recvmsg error, invali arg num_fd");
     return false;
@@ -117,7 +117,7 @@ bool vcrypto_recvmsg(int connfd, void* recv_data_buf, int recv_len, int *recv_fd
   }
 }
 
-bool vcrypto_sendmsg(int connfd, void* send_data_buf, int send_len, int send_fd, int num_fd) {
+bool vcrypto_sendmsg(int connfd, void* send_data_buf, size_t send_len, int send_fd, int num_fd) {
   if (num_fd != 1 && num_fd != 0) {
     log_error("vcrypto_sendmsg error, invalid arg num_fd");
     return false;
@@ -183,7 +183,7 @@ bool vcrypto_sendmsg(int connfd, void* send_data_buf, int send_len, int send_fd,
   }
 }
 
-bool vcrypto_recv(int connfd, void* recv_data_buf, int recv_len) {
+bool vcrypto_recv(int connfd, void* recv_data_buf, size_t recv_len) {
   int recv_len_partial = 0;
   while (recv_len > 0) {
     recv_len_partial = read(connfd, recv_data_buf, recv_len);
@@ -210,7 +210,7 @@ bool vcrypto_recv(int connfd, void* recv_data_buf, int recv_len) {
   }
 }
 
-bool vcrypto_send(int connfd, void* send_data_buf, int send_len) {
+bool vcrypto_send(int connfd, void* send_data_buf, size_t send_len) {
   int send_len_partial = 0;
   while (send_len > 0) {
     send_len_partial = write(connfd, send_data_buf, send_len);
